+++ apps/probability/distribution/poisson_distribution.h
@@ -7,7 +7,7 @@ namespace Probability {
 
 class PoissonDistribution final : public OneParameterDistribution {
 public:
-  PoissonDistribution() : OneParameterDistribution(4.0f) {}
+  PoissonDistribution() : OneParameterDistribution(4.0) {}
   I18n::Message title() override { return I18n::Message::PoissonDistribution; }
   Type type() const override { return Type::Poisson; }
   bool isContinuous() const override { return false; }
@@ -25,7 +25,7 @@ public:
   float evaluateAtAbscissa(float x) const override {
     return templatedApproximateAtAbscissa<float>(x);
   }
-  bool authorizedValueAtIndex(float x, int index) const override;
+  bool authorizedValueAtIndex(double x, int index) const override;
 private:
   double evaluateAtDiscreteAbscissa(int k) const override {
     return templatedApproximateAtAbscissa<double>(static_cast<double>(k));
